Endpoints can be exposed and Api Layer has been added
